refactor(launches): replaced filenames to index to simplify object names.
