+++ src/apps/launches/components/LaunchCard/index.css
